+++ ASDKSample/ASDKSample/BuyProductsViewController.swift
@@ -327,11 +327,11 @@ extension BuyProductsViewController: ICardScannerDelegate {
 }
 
 extension BuyProductsViewController: IRecurrentPaymentFailiureDelegate {
-    func recurrentPaymentNeedRepeatInit(additionalData: [String: String], completion: @escaping (Result<PaymentId, Error>) -> Void) {
+    func recurrentPaymentNeedRepeatInit(additionalInitData: AdditionalData, completion: @escaping (Result<PaymentId, Error>) -> Void) {
         var initData = createPaymentData()
-
-        let newPaymentData = initData.paymentFormData?.merging(additionalData) { $1 }
-        initData.paymentFormData = newPaymentData
+        var newAdditionalData = initData.additionalData ?? .empty()
+        newAdditionalData.merging(additionalInitData.data)
+        initData.additionalData = newAdditionalData
 
         coreSDK.initPayment(data: initData) { result in
             if let paymentId = try? result.map({ $0.paymentId }).get() {
@@ -617,7 +617,8 @@ private extension PaymentOptions {
         return PaymentOptions(
             orderOptions: orderOptions,
             customerOptions: customerOptions,
-            paymentData: initData.paymentFormData ?? [:]
+            paymentInitData: initData.additionalData,
+            paymentFinishData: nil
         )
     }
 }