+++ TinkoffASDKCore/TinkoffASDKCore/Domain/AcquiringAPI/Responses/Base/ASDKCoreError.swift
@@ -10,6 +10,7 @@ import Foundation
 public enum ASDKCoreError: Error {
     case missingReceiptFields
     case invalidEmail
+    case invalidFinishAuthorizeData
 }
 
 extension ASDKCoreError: LocalizedError {
@@ -20,6 +21,11 @@ extension ASDKCoreError: LocalizedError {
             return Loc.ASDKCoreError.missingReceiptFields
         case .invalidEmail:
             return Loc.ASDKCoreError.invalidEmail
+        case .invalidFinishAuthorizeData:
+            return """
+             Максимальная длина для каждого передаваемого параметра:
+             Ключ - 20 знаков. Значение - 100 знаков. Максимальное количество пар ключ:значение не может превышать 20.
+            """
         }
     }
 }