+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 
 ### Changed
 * [EACQAPW-5617] Now Receipt initializer checks validity of mandatory fields
+* [EACQAPW-5616] Fully Refactored additional DATA on /Init /FinishAuthorize and /AttachCard
 
 ### Fixed